requirements.txt
python-telegram-bot==20.8
openai==1.3.7
beautifulsoup4==4.12.2
cryptography==41.0.7
google-cloud-speech==2.22.0
requests==2.31.0
bs4==0.0.1
